internal/tmpl: add postprocessing support
This enables callers to selectively opt out of the liveview
tree model by marking a section of a template for postprocessing
using matched postprocessing start/end functions.

When postprocessing starts, we disable all tree operations,
instead writing the output directly to the writer,
which (in the case of a tree) is a buffer.
When postprocessing ends, we provide the buffer contents to
the postprocessing end function to handle as it will;
whatever it returns/writes will be recorded as a dynamic
tree element, as if the entire postprocessed section was
generated in a single operation.

Because this is an unusual need, we add no API for it.
Instead, it is hidden behind magic named functions in a funcmap.